Fix interactive remove

+++ src/app/remove.rs
@@ -13,41 +13,39 @@ impl Run for Remove {
         let mut db = DatabaseFile::new(data_dir);
         let mut db = db.open()?;
 
-        match &self.interactive {
-            Some(keywords) => {
-                let now = util::current_time()?;
-                let mut stream = db.stream(now).with_keywords(keywords);
+        if self.interactive {
+            let keywords = &self.paths;
+            let now = util::current_time()?;
+            let mut stream = db.stream(now).with_keywords(keywords);
 
-                let mut fzf = Fzf::new(true)?;
-                let stdin = fzf.stdin();
+            let mut fzf = Fzf::new(true)?;
+            let stdin = fzf.stdin();
 
-                let selection = loop {
-                    let dir = match stream.next() {
-                        Some(dir) => dir,
-                        None => break fzf.select()?,
-                    };
-
-                    match writeln!(stdin, "{}", dir.display_score(now)) {
-                        Err(e) if e.kind() == io::ErrorKind::BrokenPipe => break fzf.select()?,
-                        result => result.context("could not write to fzf")?,
-                    }
+            let selection = loop {
+                let dir = match stream.next() {
+                    Some(dir) => dir,
+                    None => break fzf.select()?,
                 };
 
-                let paths = selection.lines().filter_map(|line| line.get(5..));
-                for path in paths {
-                    if !db.remove(path) {
-                        bail!("path not found in database: {}", path);
-                    }
+                match writeln!(stdin, "{}", dir.display_score(now)) {
+                    Err(e) if e.kind() == io::ErrorKind::BrokenPipe => break fzf.select()?,
+                    result => result.context("could not write to fzf")?,
+                }
+            };
+
+            let paths = selection.lines().filter_map(|line| line.get(5..));
+            for path in paths {
+                if !db.remove(path) {
+                    bail!("path not found in database: {}", path);
                 }
             }
-            None => {
-                for path in &self.paths {
-                    if !db.remove(path) {
-                        let path_abs = util::resolve_path(path)?;
-                        let path_abs = util::path_to_str(&path_abs)?;
-                        if path_abs != path && !db.remove(path_abs) {
-                            bail!("path not found in database: {} ({})", path, path_abs)
-                        }
+        } else {
+            for path in &self.paths {
+                if !db.remove(path) {
+                    let path_abs = util::resolve_path(path)?;
+                    let path_abs = util::path_to_str(&path_abs)?;
+                    if path_abs != path && !db.remove(path_abs) {
+                        bail!("path not found in database: {} ({})", path, path_abs)
                     }
                 }
             }
